# improvement: Performance improvements to NewTag builder.