--- a/metrics/tag.go
+++ b/metrics/tag.go
@@ -6,6 +6,7 @@ package metrics
 
 import (
 	"strings"
+	"unicode"
 
 	"github.com/pkg/errors"
 )
@@ -99,15 +100,9 @@ func NewTag(k, v string) (Tag, error) {
 		return Tag{}, errors.New("value cannot be empty")
 	}
 
-	k = strings.ToLower(k)
-	if !(k[0] >= 'a' && k[0] <= 'z') {
+	firstLetter := unicode.ToLower(rune(k[0]))
+	if !(firstLetter >= 'a' && firstLetter <= 'z') {
 		return Tag{}, errors.New("tag must start with a letter")
-	}
-
-	for _, r := range k {
-		if r == ':' {
-			return Tag{}, errors.New("tag key cannot contain ':'")
-		}
 	}
 
 	// full tag, which is "key:value", must be <= 200 characters
@@ -119,8 +114,8 @@ func NewTag(k, v string) (Tag, error) {
 }
 
 func newTag(k, v string) Tag {
-	normalizedKey := normalizeTag(k)
-	normalizedValue := normalizeTag(v)
+	normalizedKey := normalizeTag(k, validKeyChars)
+	normalizedValue := normalizeTag(v, validValueChars)
 	return Tag{
 		key:      normalizedKey,
 		value:    normalizedValue,
@@ -151,20 +146,28 @@ func NewTags(t map[string]string) (Tags, error) {
 	return tags, nil
 }
 
-var validChars = make(map[rune]struct{})
+var validKeyChars = make(map[rune]struct{})
+var validValueChars = make(map[rune]struct{})
 
 func init() {
 	for ch := 'a'; ch <= 'z'; ch++ {
-		validChars[ch] = struct{}{}
+		validKeyChars[ch] = struct{}{}
+		validValueChars[ch] = struct{}{}
 	}
 	for ch := '0'; ch <= '9'; ch++ {
-		validChars[ch] = struct{}{}
+		validKeyChars[ch] = struct{}{}
+		validValueChars[ch] = struct{}{}
 	}
-	validChars['_'] = struct{}{}
-	validChars['-'] = struct{}{}
-	validChars[':'] = struct{}{}
-	validChars['.'] = struct{}{}
-	validChars['/'] = struct{}{}
+	validKeyChars['_'] = struct{}{}
+	validKeyChars['-'] = struct{}{}
+	validKeyChars['.'] = struct{}{}
+	validKeyChars['/'] = struct{}{}
+
+	validValueChars['_'] = struct{}{}
+	validValueChars['-'] = struct{}{}
+	validKeyChars[':'] = struct{}{}
+	validValueChars['.'] = struct{}{}
+	validValueChars['/'] = struct{}{}
 }
 
 // normalizeTag takes the given input string and normalizes it using the same rules as DataDog (https://help.datadoghq.com/hc/en-us/articles/204312749-Getting-started-with-tags):
@@ -173,14 +176,15 @@ func init() {
 // unicode. Tags will be converted to lowercase."
 //
 // Note that this function does not impose the length restriction described above.
-func normalizeTag(in string) string {
+func normalizeTag(in string, validChars map[rune]struct{}) string {
+	var builder strings.Builder
+	builder.Grow(len(in))
 	in = strings.ToLower(in)
-	var out string
 	for _, r := range in {
 		if _, ok := validChars[r]; !ok {
 			r = '_'
 		}
-		out += string(r)
+		_, _ = builder.WriteRune(r)
 	}
-	return out
+	return builder.String()
 }