--- a/metrics/tag.go
+++ b/metrics/tag.go
@@ -7,6 +7,7 @@ package metrics
 import (
 	"strings"
 	"unicode"
+	"unicode/utf8"
 
 	"github.com/pkg/errors"
 )
@@ -146,28 +147,30 @@ func NewTags(t map[string]string) (Tags, error) {
 	return tags, nil
 }
 
-var validKeyChars = make(map[rune]struct{})
-var validValueChars = make(map[rune]struct{})
+var validKeyChars = [utf8.RuneSelf]bool{
+	'_': true,
+	'-': true,
+	'.': true,
+	'/': true,
+}
+
+var validValueChars = [utf8.RuneSelf]bool{
+	'_': true,
+	'-': true,
+	':': true,
+	'.': true,
+	'/': true,
+}
 
 func init() {
 	for ch := 'a'; ch <= 'z'; ch++ {
-		validKeyChars[ch] = struct{}{}
-		validValueChars[ch] = struct{}{}
+		validKeyChars[ch] = true
+		validValueChars[ch] = true
 	}
 	for ch := '0'; ch <= '9'; ch++ {
-		validKeyChars[ch] = struct{}{}
-		validValueChars[ch] = struct{}{}
+		validKeyChars[ch] = true
+		validValueChars[ch] = true
 	}
-	validKeyChars['_'] = struct{}{}
-	validKeyChars['-'] = struct{}{}
-	validKeyChars['.'] = struct{}{}
-	validKeyChars['/'] = struct{}{}
-
-	validValueChars['_'] = struct{}{}
-	validValueChars['-'] = struct{}{}
-	validValueChars[':'] = struct{}{}
-	validValueChars['.'] = struct{}{}
-	validValueChars['/'] = struct{}{}
 }
 
 // normalizeTag takes the given input string and normalizes it using the same rules as DataDog (https://help.datadoghq.com/hc/en-us/articles/204312749-Getting-started-with-tags):
@@ -176,12 +179,12 @@ func init() {
 // unicode. Tags will be converted to lowercase."
 //
 // Note that this function does not impose the length restriction described above.
-func normalizeTag(in string, validChars map[rune]struct{}) string {
+func normalizeTag(in string, validChars [utf8.RuneSelf]bool) string {
 	var builder strings.Builder
 	builder.Grow(len(in))
-	in = strings.ToLower(in)
 	for _, r := range in {
-		if _, ok := validChars[r]; !ok {
+		r = unicode.ToLower(r)
+		if r >= utf8.RuneSelf || !validChars[r] {
 			r = '_'
 		}
 		_, _ = builder.WriteRune(r)