--- a/metrics/tag_test.go
+++ b/metrics/tag_test.go
@@ -1,0 +1,54 @@
+package metrics
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestMustTag(t *testing.T) {
+	t.Run("must convert uppercase to lowercase", func(t *testing.T) {
+		actual, err := NewTag("keyWithUpper", "valueWithUpper")
+		assert.NoError(t, err)
+		assert.Equal(t, Tag{
+			key:      "keywithupper",
+			value:    "valuewithupper",
+			keyValue: "keywithupper:valuewithupper",
+		},
+			actual)
+	})
+
+	t.Run("must allow special chars", func(t *testing.T) {
+		actual, err := NewTag("a_-./0", "a_-:./0")
+		assert.NoError(t, err)
+		assert.Equal(t, Tag{
+			key:      "a_-./0",
+			value:    "a_-:./0",
+			keyValue: "a_-./0:a_-:./0",
+		},
+			actual)
+	})
+
+	t.Run("must convert invalid chars", func(t *testing.T) {
+		actual, err := NewTag("a(❌)", "a(❌)")
+		assert.NoError(t, err)
+		assert.Equal(t, Tag{
+			key:      "a___",
+			value:    "a___",
+			keyValue: "a___:a___",
+		},
+			actual)
+	})
+
+	t.Run("must error when key+value empty", func(t *testing.T) {
+		_, err := NewTag("", "")
+		assert.Error(t, err)
+	})
+
+	t.Run("must error when key+value pair is longer that 200 chars", func(t *testing.T) {
+		s := strings.Repeat("a", 100)
+		_, err := NewTag(s, s)
+		assert.Error(t, err)
+	})
+}