--- a/metrics/tag_test.go
+++ b/metrics/tag_test.go
@@ -1,3 +1,7 @@
+// Copyright (c) 2021 Palantir Technologies. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package metrics
 
 import (