## Before this PR
In go 1.10 `strings.Builder` [type](https://pkg.go.dev/strings#Builder) was introduced as an effective way to build strings. 

## After this PR
Refactor the NewTag builder to use the string Builder type and optimize iterations over the key/value pairs.
==COMMIT_MSG==
Performance improvements to NewTag builder.
==COMMIT_MSG==

**Benchmarks**
```
name                 old time/op    new time/op    delta
NewTag/tagLen:2-8       156ns ± 1%     127ns ± 0%   ~     (p=0.333 n=2+2)
NewTag/tagLen:10-8      694ns ± 0%     203ns ± 3%   ~     (p=0.333 n=2+2)
NewTag/tagLen:100-8    7.90µs ± 0%    0.91µs ± 4%   ~     (p=0.333 n=2+2)
NewTag/tagLen:199-8    16.8µs ± 0%     1.7µs ± 1%   ~     (p=0.333 n=2+2)

name                 old alloc/op   new alloc/op   delta
NewTag/tagLen:2-8       12.0B ± 0%      5.0B ± 0%   ~     (p=0.333 n=2+2)
NewTag/tagLen:10-8      96.0B ± 0%     24.0B ± 0%   ~     (p=0.333 n=2+2)
NewTag/tagLen:100-8    3.54kB ± 0%    0.24kB ± 0%   ~     (p=0.333 n=2+2)
NewTag/tagLen:199-8    12.1kB ± 0%     0.4kB ± 0%   ~     (p=0.333 n=2+2)

name                 old allocs/op  new allocs/op  delta
NewTag/tagLen:2-8        3.00 ± 0%      3.00 ± 0%   ~     (all equal)
NewTag/tagLen:10-8       19.0 ± 0%       3.0 ± 0%   ~     (p=0.333 n=2+2)
NewTag/tagLen:100-8       199 ± 0%         3 ± 0%   ~     (p=0.333 n=2+2)
NewTag/tagLen:199-8       395 ± 0%         3 ± 0%   ~     (p=0.333 n=2+2
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/palantir/pkg/218)
<!-- Reviewable:end -->